# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: RailMind Backend
  version: 1.0.0
paths:
...
  /api/v1/trains:
    get:
      tags:
      - operations
      summary: Get Trains
      operationId: get_trains_api_v1_trains_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 10
          title: Page Size
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: train_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Train Type
      - name: priority
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Priority
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiListResponse_Train_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/trains/{train_id}:
    get:
      tags:
      - operations
      summary: Get Train
      operationId: get_train_api_v1_trains__train_id__get
      parameters:
      - name: train_id
        in: path
        required: true
        schema:
          type: string
          title: Train Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Train_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/train-paths:
    get:
      tags:
      - operations
      summary: Get Train Paths
      operationId: get_train_paths_api_v1_train_paths_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 10
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiListResponse_TrainPath_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/operational-windows:
    get:
      tags:
      - operations
      summary: Get Operational Windows
      operationId: get_operational_windows_api_v1_operational_windows_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 10
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiListResponse_OperationalWindow_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/train-impacts:
    get:
      tags:
      - operations
      summary: Get Train Impacts
      operationId: get_train_impacts_api_v1_train_impacts_get
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 10
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiListResponse_TrainImpact_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ApiError:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
        httpStatus:
          type: integer
          title: Httpstatus
      type: object
      required:
      - code
      - message
      - httpStatus
      title: ApiError
...
    ApiListResponse_OperationalWindow_:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/OperationalWindow'
            type: array
          - type: 'null'
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/PaginationMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiListResponse[OperationalWindow]
...
    ApiListResponse_TrainImpact_:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/TrainImpact'
            type: array
          - type: 'null'
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/PaginationMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiListResponse[TrainImpact]
...
    ApiListResponse_TrainPath_:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/TrainPath'
            type: array
          - type: 'null'
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/PaginationMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiListResponse[TrainPath]
    ApiListResponse_Train_:
      properties:
        data:
          anyOf:
          - items:
              $ref: '#/components/schemas/Train'
            type: array
          - type: 'null'
          title: Data
        pagination:
          anyOf:
          - $ref: '#/components/schemas/PaginationMeta'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiListResponse[Train]
    ApiMeta:
      properties:
        timestamp:
          type: string
          title: Timestamp
        requestId:
          type: string
          title: Requestid
        version:
          type: string
          title: Version
      type: object
      required:
      - timestamp
      - requestId
      - version
      title: ApiMeta
...
    ApiResponse_Train_:
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/Train'
          - type: 'null'
        meta:
          $ref: '#/components/schemas/ApiMeta'
        error:
          anyOf:
          - $ref: '#/components/schemas/ApiError'
          - type: 'null'
      type: object
      required:
      - meta
      title: ApiResponse[Train]
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ImpactType:
      type: string
      enum:
      - DELAY
      - CANCELLATION
      - REROUTING
      - HOLD
      - PATH_CHANGE
      title: ImpactType
...
    OperationalWindow:
      properties:
        window_id:
          type: string
          title: Window Id
        section_id:
          type: string
          title: Section Id
        interval:
          $ref: '#/components/schemas/TimeInterval'
        availability:
          $ref: '#/components/schemas/WindowAvailability'
        max_trains:
          type: integer
          title: Max Trains
        currently_assigned_trains:
          type: integer
          title: Currently Assigned Trains
      type: object
      required:
      - window_id
      - section_id
      - interval
      - availability
      - max_trains
      - currently_assigned_trains
      title: OperationalWindow
    PaginationMeta:
      properties:
        totalItems:
          type: integer
          title: Totalitems
        page:
          type: integer
          title: Page
        pageSize:
          type: integer
          title: Pagesize
        totalPages:
          type: integer
          title: Totalpages
      type: object
      required:
      - totalItems
      - page
      - pageSize
      - totalPages
      title: PaginationMeta
    PathConstraint:
      properties:
        type:
          type: string
          title: Type
        description:
          type: string
          title: Description
      type: object
      required:
      - type
      - description
      title: PathConstraint
    PathSegment:
      properties:
        section_id:
          type: string
          title: Section Id
        interval:
          $ref: '#/components/schemas/TimeInterval'
        is_conflicted:
          type: boolean
          title: Is Conflicted
      type: object
      required:
      - section_id
      - interval
      - is_conflicted
      title: PathSegment
...
    RiskLevel:
      type: string
      enum:
      - LOW
      - MEDIUM
      - HIGH
      - CRITICAL
      title: RiskLevel
...
    ScheduledTiming:
      properties:
        scheduled:
          type: string
          format: date-time
          title: Scheduled
        actual:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Actual
        delayMinutes:
          type: integer
          title: Delayminutes
      type: object
      required:
      - scheduled
      - delayMinutes
      title: ScheduledTiming
    SectionTiming:
      properties:
        section_id:
          type: string
          title: Section Id
        entry_time:
          $ref: '#/components/schemas/ScheduledTiming'
        exit_time:
          $ref: '#/components/schemas/ScheduledTiming'
      type: object
      required:
      - section_id
      - entry_time
      - exit_time
      title: SectionTiming
...
    TimeInterval:
      properties:
        start:
          type: string
          format: date-time
          title: Start
        end:
          type: string
          format: date-time
          title: End
      type: object
      required:
      - start
      - end
      title: TimeInterval
...
    Train:
      properties:
        service:
          $ref: '#/components/schemas/TrainService'
        max_speed_kmh:
          type: integer
          title: Max Speed Kmh
        length_m:
          type: integer
          title: Length M
        weight_t:
          type: integer
          title: Weight T
        priority:
          type: integer
          title: Priority
      type: object
      required:
      - service
      - max_speed_kmh
      - length_m
      - weight_t
      - priority
      title: Train
    TrainImpact:
      properties:
        train_id:
          type: string
          title: Train Id
        block_id:
          type: string
          title: Block Id
        impact_type:
          $ref: '#/components/schemas/ImpactType'
        delay_minutes:
          type: integer
          title: Delay Minutes
        is_rerouted:
          type: boolean
          title: Is Rerouted
        is_cancelled:
          type: boolean
          title: Is Cancelled
        reroute_description:
          anyOf:
          - type: string
          - type: 'null'
          title: Reroute Description
        severity:
          $ref: '#/components/schemas/RiskLevel'
        reason:
          type: string
          title: Reason
        cascading_delay_minutes:
          type: integer
          title: Cascading Delay Minutes
      type: object
      required:
      - train_id
      - block_id
      - impact_type
      - delay_minutes
      - is_rerouted
      - is_cancelled
      - severity
      - reason
      - cascading_delay_minutes
      title: TrainImpact
...
    TrainPath:
      properties:
        train_id:
          type: string
          title: Train Id
        segments:
          items:
            $ref: '#/components/schemas/PathSegment'
          type: array
          title: Segments
        constraints:
          items:
            $ref: '#/components/schemas/PathConstraint'
          type: array
          title: Constraints
        is_valid:
          type: boolean
          title: Is Valid
      type: object
      required:
      - train_id
      - segments
      - constraints
      - is_valid
      title: TrainPath
    TrainService:
      properties:
        train_id:
          type: string
          title: Train Id
        name:
          type: string
          title: Name
        train_number:
          type: string
          title: Train Number
        type:
          $ref: '#/components/schemas/TrainType'
        origin_station_id:
          type: string
          title: Origin Station Id
        destination_station_id:
          type: string
          title: Destination Station Id
        sections:
          items:
            $ref: '#/components/schemas/SectionTiming'
          type: array
          title: Sections
      type: object
      required:
      - train_id
      - name
      - train_number
      - type
      - origin_station_id
      - destination_station_id
      - sections
      title: TrainService
    TrainType:
      type: string
      enum:
      - EXPRESS
      - PASSENGER
      - FREIGHT
      - MAINTENANCE
      title: TrainType
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WindowAvailability:
      type: string
      enum:
      - AVAILABLE
      - OCCUPIED
      - MAINTENANCE
      - CLOSED
      title: WindowAvailability
...